Improve HUD font: add Latin-1 Supplement and other missing chars  * Adds things like custom accented chars, #, &, $, ~, etc  * Still missing some Latin-1 Supplement characters (copyright, paragraph, and whatnot)  * Allows for non-broken Spanish/Portuguese translations (see #10, #8)
Note: Due to the added accented chars, the Win Ascent was bumped up so that the accents do not cause the characters to be clipped. This alters the default line height of the font, which makes the text much smaller in-game. To compensate, the font sizes should probably be bumped up in the Scheme.res files.